Explore suitability of Python's generators
After a brief false hope, I have come to the conclusion that I can't get
everything I want from Python's generators. There's no way to 'replay'
the stream from an earlier point.

This is a shame, because I liked the idea of using 'yield' to offer up
values in a stream from functions written in straightforward imperative
styles. I think this would make it easier to introduce the abstraction
to others, as opposed to the somewhat convoluted
recursive-closure-trampoline syntax I worked with previously.

It may be worth exploring a generator version a little further anyway,
for that reason. I should also see if I can come up with compelling use
cases for replayability, beyond repeating a finite stream (which could
also be achieved less elegantly with generator factories).