Text-to-speech UI; revamp list->stream, pickling.
list_to_stream dropped the reference to the list and was impossible for
any pickle-derivative to serialize, because the list lengths easily
exceed the max recursion depth.

Now, list_to_stream has been replaced by ListStream, which is easier to
serialize, easier to inspect (includes length), and slightly faster.

For now, switch to cloudpickle because it can pickle streams directly.
(Tried dill first, but it failed even on `osc(440)`.)