Fix types for fauxdot optional dependency.
As before, we want to defer the import error to first usage.
Previously, pylance & co. got the types wrong because of this,
and would mark code after calls to beat/tune as unreachable.
Using `typing.TYPE_CHECKING`, which is always False at runtime but
considered True by type-checkers, allows us to avoid this.